+++ jenkins/Jenkinsfile
@@ -17,6 +17,7 @@ pipeline {
                 sh "docker build --target build -t opencue/pyoutline-build:${env.BUILD_ID} -f pyoutline/Dockerfile ."
                 sh "docker build --target build -t opencue/cuegui-build:${env.BUILD_ID} -f cuegui/Dockerfile --build-arg=upstream_version=${env.BUILD_ID} ."
                 sh "docker build --target build -t opencue/cuesubmit-build:${env.BUILD_ID} -f cuesubmit/Dockerfile ."
+                sh "docker build --target build -t opencue/cueadmin-build:${env.BUILD_ID} -f cueadmin/Dockerfile ."
             }
         }
         stage('Test') {
@@ -27,6 +28,7 @@ pipeline {
                 sh "docker build --target test -t opencue/pyoutline-test:${env.BUILD_ID} -f pyoutline/Dockerfile ."
                 sh "docker build --target test -t opencue/cuegui-test:${env.BUILD_ID} -f cuegui/Dockerfile --build-arg=upstream_version=${env.BUILD_ID} ."
                 sh "docker build --target test -t opencue/cuesubmit-test:${env.BUILD_ID} -f cuesubmit/Dockerfile ."
+                sh "docker build --target test -t opencue/cueadmin-test:${env.BUILD_ID} -f cueadmin/Dockerfile ."
             }
         }
         stage('Package') {
@@ -37,6 +39,7 @@ pipeline {
                 sh "docker build -t opencue/pyoutline:${env.BUILD_ID} -f pyoutline/Dockerfile ."
                 sh "docker build -t opencue/cuegui:${env.BUILD_ID} -f cuegui/Dockerfile --build-arg=upstream_version=${env.BUILD_ID} ."
                 sh "docker build -t opencue/cuesubmit:${env.BUILD_ID} -f cuesubmit/Dockerfile ."
+                sh "docker build -t opencue/cueadmin:${env.BUILD_ID} -f cueadmin/Dockerfile ."
                 sh "jenkins/extract_artifacts.sh ${env.BUILD_ID} ./build"
             }
         }
@@ -54,6 +57,7 @@ pipeline {
             archiveArtifacts artifacts: 'build/pyoutline-*-all.tar.gz', fingerprint: true
             archiveArtifacts artifacts: 'build/cuegui-*-all.tar.gz', fingerprint: true
             archiveArtifacts artifacts: 'build/cuesubmit-*-all.tar.gz', fingerprint: true
+            archiveArtifacts artifacts: 'build/cueadmin-*-all.tar.gz', fingerprint: true
         }
     }
 }
